+++ lib/fireap.rb
@@ -1,4 +1,8 @@
 module Fireap
+  NAME = 'fireap'
+  module Kv
+    PREFIX = Fireap::NAME + '/'
+  end
   class Error < StandardError
   end
 end
