+++ lib/fireap/data/kv.rb
@@ -1,0 +1,24 @@
+require 'base64'
+
+require 'fireap/model/kv'
+
+module Fireap::Data
+  class Kv
+    @me = nil
+    def initialize(kv)
+      @me = kv
+    end
+
+    def to_data
+      data = {}
+      @me.each do |key, val|
+        if key == 'Value'
+          data[key.to_snakecase] = Base64.decode64(val)
+        else
+          data[key.to_snakecase] = val
+        end
+      end
+      Fireap::Model::Kv.new(data)
+    end
+  end
+end
